Moved python scripts to separate directory

+++ python/centrality.py
@@ -1,9 +1,11 @@
 # coding: utf-8
 
 from graph_tool.all import *
+import numpy as np
+import scipy as sp
 
 def cen_degree(g):
-    return g.degree_property_map("total").a / (g.num_vertices() - 1)
+    return g.degree_property_map("total").a.astype(float) / (g.num_vertices() - 1)
 
 def cen_katz(g):
     return katz(g).a
@@ -21,5 +23,5 @@ def nlargest_idx(a, n):
 def write_nodes(a, fname):
     f = open(fname, 'w')
     for i in a:
-        f.write("%d\n" % (i+1))
+        f.write("%d\n" % (i+1)) # convert from 0-based index to 1-based
     f.close()
